+++ src/main/java/com/br/medtech/service/DoctorService.java
@@ -17,7 +17,7 @@ public class DoctorService {
     @Autowired
     private DoctorRepository doctorRepository;
 
-    public ResponseEntity<Optional<Doctor>> findDoctorById(Long id) {
+    public ResponseEntity<Optional<Doctor>> findDoctorById(Integer id) {
         try {
             Optional<Doctor> doctor = doctorRepository.findById(id);
 
@@ -41,7 +41,7 @@ public class DoctorService {
         }
     }
 
-    public ResponseEntity<Doctor> update(Long id, Doctor doctor) {
+    public ResponseEntity<Doctor> update(Integer id, Doctor doctor) {
         if (!doctorRepository.existsById(id)) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Médico não encontrado no banco de dados.");
         }
@@ -55,7 +55,7 @@ public class DoctorService {
         }
     }
 
-    public ResponseEntity<Doctor> delete(Long id){
+    public ResponseEntity<Doctor> delete(Integer id){
         if (!doctorRepository.existsById(id)) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Médico não encontrado no banco de dados.");
         }
